Fix the other concurrency issue w.r.t to the Map of ArtifactDependencyManager.
I moved the Map within the ArtifactDependencyManager under the synchronized lock.
I believe the plugin should now work for parallel builds as well as sets of projects
with multiple invocations of the plugin. (Although the plugin is really only needed
in the APK projects, it can be used in other projects to flag issues within those
projects.