Downgrade progress message from warning to info level
Any output from a Gradle task causes the task line, the message, and
a blank line to be added to the output; whereas tasks with no output
are "wiped" or overwritten. This "noise" in the output competes with
real warnings, and other messages that we shouldn't ignore. Anything
that isn't a real warning shouldn't be output at this logging level.

This moves the "Parsing json file" message into the info level, which
is more appropriate for a status message, which carries no corrective
action for the developer.