Add special handling for library projects.
Library projects are ambiguous when attempting to resolve their dependencies. Internally there
are many variants and Gradle will throw a AmbiguousVariantSelectionException when it encounters
this scenario. The result is that a null set of artifacts is returned from getResolvedArtifacts().

This change iterates through the dependency hierarchy's top levels manually to avoid attempting
to automatically traverse a library project. Instead it will manually traverse the tree of library
projects and only automatically traverse non-library dependencies.