Make DependencyTask cacheable/incremental
Checking all build.gradle files in the project as input key,
assuming that you have to change dependencies in build.gradle
to have a change in the generated/used licenses.
As output only the single file is declared,
because on the output directory the exclusive access
(for caching required, see: https://guides.gradle.org/using-build-cache/#concepts_overlapping_outputs)
can't be guaranteed.