Use a Set rather than a List to store dependencies in ArtifactDependencyManager
Using a List means any dependency encountered more than once will be stored multiple times which unnecessarily consumes memory. The class API returns a Set for the dependency list, so using a Set internally does not alter the behavior of the class whilst giving a memory saving on dependency graphs where the same relationship might be encountered twice

I've added a test which demonstrates the duplicate use problem. With the existing code this test will fail due to storing multiple copies of the same relationship.