Fix incorrect wiring of the GoogleServicesTask
In the process of making configuration of the GooglerServicesTask lazily
configured, the registration of the generated outputs with the Android
Gradle plugin was moved into the lazy task configuration. This only
works if the google services task is configured before any of the tasks
that consume it, which is not guaranteed, leading to the below issues.

This change also removes the partial workaround of manually adding the
task dependency to one of the consuming tasks.

Issue: https://github.com/google/play-services-plugins/issues/228
Also potentially related: https://github.com/google/play-services-plugins/issues/228
Test: Manually verified that the wiring works correctly with AGP 7.4.0-alpha03 and Gradle 7.5-rc-1