Add google-services.json as a task input, to fix up-to-date checks
`google-services.json` was not previously treated as a task input,
so changes to that file would not cause the task's UP-TO-DATE check
to return `false`. As a result, the generated `values.xml` file is
never updated unless `clean` is run after editing the json file, which
seems buggy and has customer problems.

See [this issue](https://github.com/firebase/firebase-android-sdk/issues/2191)
in the firebase-android-sdk repo, as an example.